fix(instance list): Fixes
- Fixed issues with whatsapp status not changing on instance list
