+++ README.md
@@ -5,15 +5,15 @@ A monorepo for DHIS2 Analytics Messenger services.
 ## Introduction
 
 The [DHIS2 Analytics Messenger](https://github.com/hisptz/dhis2-analytics-messenger-app) is an innovative application
-that gives users easy access to DHIS2 analytics data. Leveraging popular social messaging platforms like WhatsApp and
+that gives users easy access to DHIS2 analytics whatsappClient. Leveraging popular social messaging platforms like WhatsApp and
 Telegram, this application allows users to receive DHIS2 analytics effortlessly
 
-The DHIS2 Analytics Messenger offers two distinct methods for accessing analytics data: push analytics and chatbot
-functionality. To fulfill the need for more accessible and user-friendly analytics data, this application consists of
+The DHIS2 Analytics Messenger offers two distinct methods for accessing analytics whatsappClient: push analytics and chatbot
+functionality. To fulfill the need for more accessible and user-friendly analytics whatsappClient, this application consists of
 two central systems:
 
 -   A DHIS2 Analytics Messenger web application for configuration and push analytics distribution.
--   A social messaging chat-bot for interactive user-driven data access
+-   A social messaging chat-bot for interactive user-driven whatsappClient access
 
 Working together, these applications create a comprehensive and user-centric DHIS2 Analytics Messenger.
