+++ services/core/.gitignore
@@ -11,7 +11,7 @@ lerna-debug.log*
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
-# Runtime data
+# Runtime whatsappClient
 pids
 *.pid
 *.seed
