+++ services/messaging/.gitignore
@@ -10,7 +10,7 @@ lerna-debug.log*
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
-# Runtime data
+# Runtime whatsappClient
 pids
 *.pid
 *.seed
